scraping content changed disclaimer
its impossible to keep this constantly up-to-date. scraping will always change with content of web page you are trying to scrape. If someone can deploy a  static site that we maintain and use it for this readying, that would be good. Otherwise i think the disclaimer is okay for now because this is a reading, not a code-along.